+++ lib/AmazonApps/Common/src/index.ts
@@ -1,0 +1,2 @@
+export * from './AAwsApplication';
+export * from './AAwsObjectConnector';