+++ lib/AmazonApps/Redshift/src/RedshiftApplication.ts
@@ -4,7 +4,8 @@ import {
     DescribeClustersCommandInput,
     RedshiftClient,
 } from '@aws-sdk/client-redshift';
-import AAwsApplication, {
+import {
+    AAwsApplication,
     CREDENTIALS,
     KEY,
     LATEST,
@@ -12,7 +13,7 @@ import AAwsApplication, {
     REGIONS,
     SECRET,
     VERSION,
-} from '@orchesty/connector-amazon-apps-common/src/AAwsApplication';
+} from '@orchesty/connector-amazon-apps-common';
 import CoreFormsEnum, { getFormName } from '@orchesty/nodejs-sdk/dist/lib/Application/Base/CoreFormsEnum';
 import { ApplicationInstall } from '@orchesty/nodejs-sdk/dist/lib/Application/Database/ApplicationInstall';
 import Field from '@orchesty/nodejs-sdk/dist/lib/Application/Model/Form/Field';